requirements.txt
langchain-google-genai
python-dotenv
scikit-learn
numpy
